requirements.txt
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
